update-feature-fix-bugs
load data to manage promotions view, install popup for list when selecting an item, upload data from excel file and save it to database, fix some bug in BUS, DAO

+++ Project1_BookStore/DTO/PromotionDTO.cs
@@ -10,7 +10,7 @@ namespace Project1_BookStore.DTO
     {
         public string promoID { get; set; }
         public string promoName { get; set; }
-        public double promoDiscount { get; set; }
+        public float promoDiscount { get; set; }
         public string promoDesciption { get; set; }
         public DateTime promoStartTime { get; set; }
         public DateTime promoEndTime { get; set; }
